refactor: replace thunks and rename reducers
