+++ src/redux/reducers/taskTrackPageIndex.js
@@ -1,6 +1,6 @@
 import firebaseApi from '../../api/firebaseApi';
 import dateToStringForInput from '../../helpers/dateToStringForInput';
-import { toggleIsFetching } from './appReducer';
+import { toggleIsFetching } from './appIndex';
 
 const SHOW_TASK_TRACK_PAGE = 'SHOW_TASK_TRACK_PAGE';
 const SET_TASK_TRACK = 'SET_TASK_TRACK';
@@ -13,21 +13,6 @@ export const showTaskTrackPage = (taskTrackPageIsVisible) => ({ type: SHOW_TASK_
 export const setTrackNote = (trackNote) => ({ type: SET_TASK_TRACK, trackNote });
 export const onChangeValue = (fieldName, value) => ({ type: ON_CHANGE, [fieldName]: value });
 export const clearTaskTrackPage = () => ({ type: CLEAR_TASK_TRACK_PAGE });
-
-export const getTaskTrack = (taskTrackId) => (dispatch) => {
-  dispatch(toggleIsFetching(true));
-  firebaseApi
-    .getTaskTrack(taskTrackId)
-    .then((result) => {
-      if (result.message) {
-        dispatch(setError(result));
-      }
-      dispatch(setTrackNote(result));
-    })
-    .then(() => {
-      dispatch(toggleIsFetching(false));
-    });
-};
 
 const initialState = {
   taskTrackPageIsVisible: false,
@@ -55,6 +40,21 @@ const taskTrackPageReducer = (state = initialState, action) => {
     default:
       return state;
   }
+};
+
+export const getTaskTrack = (taskTrackId) => (dispatch) => {
+  dispatch(toggleIsFetching(true));
+  firebaseApi
+    .getTaskTrack(taskTrackId)
+    .then((result) => {
+      if (result.message) {
+        dispatch(setError(result));
+      }
+      dispatch(setTrackNote(result));
+    })
+    .then(() => {
+      dispatch(toggleIsFetching(false));
+    });
 };
 
 export default taskTrackPageReducer;
